pyproject.toml
[tool.poetry]
name = "emoji_injection_proxy_server"
version = "0.1.0"
description = ""
authors = ["Victor Luzgin"]

[tool.poetry.dependencies]
python = "^3.8"
"proxy.py" = "^2.3.1"
beautifulsoup4 = "^4.9.3"
emoji = "^1.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
